requirements.txt
psycopg2-binary==2.9.9
sqlparse==0.4.4
pandas==2.0.3
tqdm==4.66.1
docker==6.1.3
openpyxl==3.1.2
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
tabulate==0.9.0
